Break progress bar out of editor component
